+++ backend/Kanbas/Modules/routes.js
@@ -3,12 +3,22 @@ import * as dao from "./dao.js";
 
 export default function ModuleRoutes(app) {
     const createModule = async (req, res) => {
+        const currentUser = req.session["currentUser"];
+        if (!currentUser || currentUser.role != "FACULTY") {
+            res.sendStatus(401);
+            return;
+        }
         const { cid } = req.params;
         const module = { ...req.body, course: cid }
         const newModule = await dao.createModule(module)
         res.json(newModule)
     }
     const deleteModule = async (req, res) => {
+        const currentUser = req.session["currentUser"];
+        if (!currentUser || currentUser.role != "FACULTY") {
+            res.sendStatus(401);
+            return;
+        }
         const status = await dao.deleteModule(req.params.mid)
         res.json(status)
     }
@@ -18,6 +28,11 @@ export default function ModuleRoutes(app) {
         res.send(modules)
     }
     const updateModule = async (req, res) => {
+        const currentUser = req.session["currentUser"];
+        if (!currentUser || currentUser.role != "FACULTY") {
+            res.sendStatus(401);
+            return;
+        }
         const { mid } = req.params;
         const status = await dao.updateModule(mid, req.body);
         res.json(status);
@@ -27,5 +42,4 @@ export default function ModuleRoutes(app) {
     app.delete("/api/modules/:mid", deleteModule)
     app.put("/api/modules/:mid", updateModule);
     app.get("/api/courses/:cid/modules", findModulesByCourse);
-
 }
